Update `toMatch` to `toMatchText`
Hey @neoziro, thanks so much for setting up this repo. It looks like it
has a lot of valuable content in it to get expectations up and running
quickly with Puppeteer and Jest.

I was hoping you could take a look at this proposed change for matching
text. We were doing something similar in one of our projects and we
started looking around for better ways of matching text. It turns out
that @aslushnikov proposed a nice quick way of doing so [here](https://github.com/GoogleChrome/puppeteer/issues/438#issuecomment-340675046).

It does change the api a bit, so it'd be a breaking change. I think in
the end it is worth it though, as using a regex to search the entire
body of the DOM can be a large computation. Using a selector will help
minimize that computation and also find that result in a way that the
puppeteer team has suggested.

Let me know your thoughts! Hope to hear from you soon.